requirements (2).txt
Flask==3.0.3
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
gunicorn==22.0.0
google-cloud-language==2.16.0
vertexai==1.71.1